qlib_crypto_trading/requirements.txt
# Qlib for quantitative research
pyqlib>=0.9.0

# Core data manipulation
pandas>=1.4.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Deep learning (optional, for neural network models)
torch>=1.10.0
tensorflow>=2.8.0

# Cryptocurrency exchange APIs
ccxt>=1.70.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
loguru>=0.6.0
redis>=4.0.0

# Data processing and storage
pyarrow>=6.0.0
h5py>=3.6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0

# Async programming
asyncio-mqtt>=0.11.0
websockets>=10.0